Final MVC refactor
- Renamed and moved SimulatorView to CarparkHandler
- Removed 100 step, too buggy
- Simulator is now independent from views
- updateViews() is working and being used
