+++ src/nl/simulator/mvc/controller/Controller.java
@@ -11,7 +11,6 @@ import java.awt.event.*;
 public class Controller extends AbstractController implements ActionListener{
     private Simulator simulator;
     private JButton step1;
-    private JButton step100;
     private JButton start;
     private JButton stop;
 
@@ -20,16 +19,13 @@ public class Controller extends AbstractController implements ActionListener{
         this.simulator = simulator;
 
         step1 = new JButton("1 step");
-        step100 = new JButton("100 steps");
         start = new JButton("Start");
         stop = new JButton("Stop");
         step1.addActionListener(this);
-        step100.addActionListener(this);
         start.addActionListener(this);
         stop.addActionListener(this);
 
         this.add(step1);
-        this.add(step100);
         this.add(start);
         this.add(stop);
 
@@ -40,8 +36,6 @@ public class Controller extends AbstractController implements ActionListener{
     public void actionPerformed(ActionEvent e) {
         if(e.getSource() == step1) {
             simulator.run(1);
-        } else if(e.getSource() == step100) {
-            simulator.run(100);
         } else if(e.getSource() == start) {
             simulator.start();
         } else if(e.getSource() == stop) {
